Made everything into one app
