--- backend/appointment/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from .models import Appointment
-
-@admin.register(Appointment)
-class AppointmentAdmin(admin.ModelAdmin):
-    list_display = ('patient', 'doctor', 'schedule', 'status', 'created_at')
-    list_filter = ('status', 'doctor')
-    search_fields = ('patient__first_name', 'patient__last_name', 'doctor__first_name', 'doctor__last_name')
