--- backend/appointment/models.py
@@ -1,21 +1,0 @@
-from django.db import models
-from doctor.models import Doctor, DoctorSchedule
-from patient.models import Patient
-
-class Appointment(models.Model):
-    STATUS_CHOICES = [
-        ('scheduled', 'Scheduled'),
-        ('completed', 'Completed'),
-        ('canceled', 'Canceled'),
-        ('no_show', 'No Show'),
-    ]
-    
-    patient = models.ForeignKey(Patient, on_delete=models.CASCADE, related_name='appointments')
-    doctor = models.ForeignKey(Doctor, on_delete=models.CASCADE, related_name='appointments')
-    schedule = models.ForeignKey(DoctorSchedule, on_delete=models.CASCADE, related_name='appointments')
-    status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='scheduled')
-    notes = models.TextField(blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    
-    def __str__(self):
-        return f"{self.patient} with {self.doctor} on {self.schedule.date} at {self.schedule.start_time}"
