--- backend/doctor/views.py
@@ -1,94 +1,0 @@
-from rest_framework import viewsets, generics
-from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from .models import Doctor, DoctorSchedule, Specialty, Language
-from .serializers import (
-    DoctorSerializer, DoctorScheduleSerializer, 
-    SpecialtySerializer, LanguageSerializer
-)
-
-class SpecialtyViewSet(viewsets.ModelViewSet):  # Changed from ReadOnlyModelViewSet
-    queryset = Specialty.objects.all()
-    serializer_class = SpecialtySerializer
-    
-    def get_permissions(self):
-        """Read access for everyone, write access for admins only"""
-        if self.action in ['list', 'retrieve']:
-            permission_classes = [AllowAny]
-        else:
-            permission_classes = [IsAdminUser]
-        return [permission() for permission in permission_classes]
-
-class LanguageViewSet(viewsets.ModelViewSet):  # Changed from ReadOnlyModelViewSet
-    queryset = Language.objects.all()
-    serializer_class = LanguageSerializer
-    
-    def get_permissions(self):
-        """Read access for everyone, write access for admins only"""
-        if self.action in ['list', 'retrieve']:
-            permission_classes = [AllowAny]
-        else:
-            permission_classes = [IsAdminUser]
-        return [permission() for permission in permission_classes]
-
-class DoctorViewSet(viewsets.ModelViewSet):  # Changed from ReadOnlyModelViewSet
-    queryset = Doctor.objects.all()
-    serializer_class = DoctorSerializer
-    filterset_fields = ['specialties', 'languages']
-    search_fields = ['first_name', 'last_name', 'bio']
-    
-    def get_permissions(self):
-        """
-        Different permissions for different actions:
-        - Read operations (list, retrieve): Anyone can access
-        - Write operations (create, update, delete): Admin only
-        """
-        if self.action in ['list', 'retrieve']:
-            permission_classes = [AllowAny]
-        else:  # create, update, partial_update, destroy
-            permission_classes = [IsAdminUser]
-        return [permission() for permission in permission_classes]
-
-class DoctorsBySpecialtyView(generics.ListAPIView):
-    serializer_class = DoctorSerializer
-    permission_classes = [AllowAny]
-    
-    def get_queryset(self):
-        specialty_id = self.kwargs['specialty_id']
-        return Doctor.objects.filter(specialties__id=specialty_id)
-
-class DoctorScheduleViewSet(viewsets.ModelViewSet):
-    queryset = DoctorSchedule.objects.all()
-    serializer_class = DoctorScheduleSerializer
-    
-    def get_permissions(self):
-        if self.action in ['list', 'retrieve']:
-            permission_classes = [AllowAny]
-        else:
-            permission_classes = [IsAdminUser]
-        return [permission() for permission in permission_classes]
-
-class AvailableSlotsView(APIView):
-    permission_classes = [AllowAny]
-    
-    def get(self, request, doctor_id):
-        from datetime import datetime, timedelta
-        
-        # Get today's date
-        today = datetime.now().date()
-        # Get end date (30 days from now by default)
-        doctor = Doctor.objects.get(id=doctor_id)
-        end_date = today + timedelta(days=30)
-        
-        # Get available schedules
-        available_slots = DoctorSchedule.objects.filter(
-            doctor=doctor,
-            date__gte=today,
-            date__lte=end_date,
-            is_active=True,
-            available_slots__gt=0
-        ).order_by('date', 'start_time')
-        
-        serializer = DoctorScheduleSerializer(available_slots, many=True)
-        return Response(serializer.data)
