--- backend/patient/models.py
@@ -1,32 +1,0 @@
-from django.db import models
-from django.conf import settings
-
-class Patient(models.Model):
-    user = models.OneToOneField(
-        settings.AUTH_USER_MODEL, 
-        on_delete=models.CASCADE, 
-        null=True, 
-        blank=True,
-        related_name='patient_profile'
-    )
-    first_name = models.CharField(max_length=100)
-    last_name = models.CharField(max_length=100)
-    date_of_birth = models.DateField()
-    phone_number = models.CharField(max_length=15)
-    address = models.TextField()
-    
-    emergency_contact_name = models.CharField(max_length=100, blank=True)
-    emergency_contact_number = models.CharField(max_length=15, blank=True)
-    
-    def __str__(self):
-        return f"{self.first_name} {self.last_name}"
-
-class MedicalHistory(models.Model):
-    patient = models.ForeignKey(Patient, on_delete=models.CASCADE, related_name='medical_histories')
-    diagnosis = models.CharField(max_length=200)
-    diagnosis_date = models.DateField()
-    treatment = models.TextField()
-    notes = models.TextField(blank=True)
-    
-    def __str__(self):
-        return f"{self.diagnosis} - {self.patient}"
