--- backend/patient/serializers.py
@@ -1,12 +1,0 @@
-from rest_framework import serializers
-from .models import Patient, MedicalHistory
-
-class PatientSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Patient
-        fields = '__all__'
-
-class MedicalHistorySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = MedicalHistory
-        fields = '__all__'
