--- backend/patient/views.py
@@ -1,28 +1,0 @@
-from rest_framework import viewsets
-from rest_framework.permissions import IsAuthenticated
-from django.shortcuts import get_object_or_404
-from .models import Patient, MedicalHistory
-from .serializers import PatientSerializer, MedicalHistorySerializer
-
-class PatientViewSet(viewsets.ModelViewSet):
-    serializer_class = PatientSerializer
-    permission_classes = [IsAuthenticated]
-    
-    def get_queryset(self):
-        # Users can only see their own patient profile
-        return Patient.objects.filter(user=self.request.user)
-    
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
-
-class MedicalHistoryViewSet(viewsets.ModelViewSet):
-    serializer_class = MedicalHistorySerializer
-    permission_classes = [IsAuthenticated]
-    
-    def get_queryset(self):
-        # Users can only see their own medical history
-        return MedicalHistory.objects.filter(patient__user=self.request.user)
-    
-    def perform_create(self, serializer):
-        patient = get_object_or_404(Patient, user=self.request.user)
-        serializer.save(patient=patient)
